+++ crates/cachet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2026-04-20
+
+- ⚠️ Breaking
+
+  - bump `ohno` to 0.4.0 and `cachet_tier` to 0.2.0
+
 ## [0.1.0]
 
 Initial release.